feat: migrate /firestation endpoint to AlertInfoController
- Migrated getPersonsCoveredByStation method from FireStationController to AlertInfoController
- Moved related business logic from FireStationService to AlertInfoService
- Verified endpoint functionality at /firestation?stationNumber=<stationNumber> via Postman
