+++ src/main/java/com/openclassrooms/safetynet/safetynetapi/controller/AlertInfoController.java
@@ -1,6 +1,7 @@
 package com.openclassrooms.safetynet.safetynetapi.controller;
 
 import com.openclassrooms.safetynet.safetynetapi.dto.ChildDTO;
+import com.openclassrooms.safetynet.safetynetapi.dto.CoveredPersonsByStationDTO;
 import com.openclassrooms.safetynet.safetynetapi.dto.PersonInfoDto;
 import com.openclassrooms.safetynet.safetynetapi.service.AlertInfoService;
 import lombok.extern.log4j.Log4j2;
@@ -107,5 +108,24 @@ public class AlertInfoController {
         log.debug("Number of phone numbers found: {}", phoneNumbers.size());
         return ResponseEntity.ok(phoneNumbers);
 
+    }
+
+    /**
+     * GET endpoint that retrieves a list of persons covered by a given fire station number.
+     * <p>
+     * The response includes:
+     * <ul>
+     *     <li>Basic personal information (first name, last name, address, phone)</li>
+     *     <li>The number of adults and children among the listed persons</li>
+     * </ul>
+     *
+     * @param stationNumber the fire station number used to filter addresses and retrieve associated persons
+     * @return a ResponseEntity containing a CoveredPersonsByStationDTO with the list of persons and counts,
+     * or an appropriate HTTP error if no data is found
+     */
+    @GetMapping("/firestation")
+    public ResponseEntity<CoveredPersonsByStationDTO> getPersonsCoveredByStation(@RequestParam("stationNumber") int stationNumber) {
+        CoveredPersonsByStationDTO response = alertInfoService.getPersonsCoveredByStation(stationNumber);
+        return ResponseEntity.ok(response);
     }
 }
